move sharedv4 test to it's own file
