+++ tests/basic/sharedv4_svc_test.go
@@ -4,15 +4,12 @@ import (
 	"fmt"
 	"os"
 	"path"
-	"regexp"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/libopenstorage/openstorage/api"
-	"github.com/libopenstorage/openstorage/pkg/mount"
-	"github.com/portworx/sched-ops/k8s/apps"
 	"github.com/portworx/sched-ops/k8s/core"
 	"github.com/portworx/sched-ops/k8s/errors"
 	"github.com/portworx/torpedo/drivers/node"
@@ -36,7 +33,6 @@ const (
 	testConnTimeout      = 15 * time.Minute
 	offlineClientTimeout = 15 * time.Minute // as defined in ref.go
 	exportPathPrefix     = "/var/lib/osd/pxns"
-	devicePathPrefix     = "/dev/pxd/pxd"
 )
 
 // Verify that the pods on old and new NFS servers get terminated after a failover.
@@ -331,7 +327,7 @@ var _ = Describe("{Sharedv4SvcFunctional}", func() {
 		testSv4Contexts = getTestSv4Contexts(contexts)
 		testSharedV4Contexts = getTestSharedV4Contexts(contexts)
 		if len(testSv4Contexts) == 0 && len(testSharedV4Contexts) == 0 {
-			Skip("No test-sv4-svc or test-sharedv4 apps were found")
+			Skip("No test-sv4-svc apps were found")
 		}
 		workers = node.GetWorkerNodes()
 		numPods = len(workers)
@@ -1060,121 +1056,6 @@ var _ = Describe("{Sharedv4SvcFunctional}", func() {
 			}
 		})
 
-		Context("{SharedV4MountRecovery}", func() {
-			BeforeEach(func() {
-				namespacePrefix = "sharedv4mountrecovery"
-			})
-
-			JustBeforeEach(func() {
-				for _, ctx := range testSharedV4Contexts {
-					vol, _, attachedNode := getSv4TestAppVol(ctx)
-					k8sApps := apps.Instance()
-					Step(
-						fmt.Sprintf("setup app %s to deploy on single node", ctx.App.Key),
-						func() {
-							Step(fmt.Sprintf("add label to attached node for %s", ctx.App.Key), func() {
-								err := core.Instance().AddLabelOnNode(attachedNode.Name, "attachedNode", "true")
-								Expect(err).NotTo(HaveOccurred())
-							})
-
-							Step(fmt.Sprintf("update %s deployment with label", ctx.App.Key), func() {
-								deployments, err := k8sApps.ListDeployments(ctx.GetID(), metav1.ListOptions{})
-								Expect(err).NotTo(HaveOccurred())
-								deployment := deployments.Items[0]
-								deployment.Spec.Template.Spec.NodeSelector = map[string]string{"attachedNode": "true"}
-								deployment.Spec.Template.Spec.Affinity = nil
-
-								_, err = k8sApps.UpdateDeployment(&deployment)
-								Expect(err).NotTo(HaveOccurred())
-							})
-
-							Step(fmt.Sprintf("scale down app: %s to 0 ", ctx.App.Key), func() {
-								scaleApp(ctx, 0)
-							})
-
-							// wait until all pods are gone
-							Step(fmt.Sprintf("wait for app %s to have 0 pods", ctx.App.Key), func() {
-								waitForNumPodsToEqual(ctx, 0)
-							})
-
-							Step(fmt.Sprintf("scale it backsssss %s", ctx.App.Key), func() {
-								scaleApp(ctx, numPods)
-
-								ValidateContext(ctx)
-							})
-
-							// validate the pods are all on one node
-							Step(fmt.Sprintf("validate that all pods are running on the attached node for %s", ctx.App.Key), func() {
-								pods, err := core.Instance().GetPodsUsingPV(vol.ID)
-								Expect(err).NotTo(HaveOccurred())
-								for _, pod := range pods {
-									Expect(pod.Spec.NodeName).To(Equal(attachedNode.Name))
-								}
-
-							})
-						})
-				}
-
-			})
-
-			It("should set device path to RO and validate recovery", func() {
-				for _, ctx := range testSharedV4Contexts {
-					_, apiVol, attachedNode := getSv4TestAppVol(ctx)
-					counterCollectionInterval := 3 * time.Duration(numPods) * time.Second
-					devicePath := fmt.Sprintf("%s%s", devicePathPrefix, apiVol.Id)
-
-					// In porx, we check the export path is read-only when it is previously RO and
-					// changed to RW by fs due to occuring error. Specifically, when mnt.Opts (default)
-					// is RW and mnt.VfsOpts (fs state) is RO.
-					// By setting device path to RO, we mock the same behavior. Check state in
-					// `/proc/self/mountinfo`
-					Step(fmt.Sprintf("mark device path as RO %s", ctx.App.Key), func() {
-						setPathToROMode(devicePath, attachedNode)
-					})
-
-					Step(fmt.Sprintf("validate the counters are inactive %s", ctx.App.Key), func() {
-						counters := getAppCounters(apiVol, attachedNode, counterCollectionInterval)
-						activePods := getActivePods(counters)
-						Expect(len(activePods)).To(Equal(0))
-					})
-
-					Step(fmt.Sprintf("restart vol driver %s", ctx.App.Key), func() {
-						restartVolumeDriverOnNode(attachedNode)
-					})
-
-					Step(fmt.Sprintf("validate counter are active for %s", ctx.App.Key), func() {
-						counters := getAppCounters(apiVol, attachedNode, counterCollectionInterval)
-						activePods := getActivePods(counters)
-						Expect(len(activePods)).To(Equal(numPods))
-					})
-
-					Step(fmt.Sprintf("validate device path is set as RW for %s", ctx.App.Key), func() {
-						mntList, err := mount.GetMounts()
-						Expect(err).NotTo(HaveOccurred())
-						var volumeMountRW = regexp.MustCompile(`,rw,|,rw|rw,|rw`)
-
-						for _, mnt := range mntList {
-							if mnt.Mountpoint != devicePath {
-								continue
-							}
-							Expect(volumeMountRW.MatchString(mnt.Opts)).To(BeTrue())
-							Expect(volumeMountRW.MatchString(mnt.VfsOpts)).To(BeTrue())
-						}
-					})
-				}
-			})
-
-			JustAfterEach(func() {
-				for _, ctx := range testSharedV4Contexts {
-					_, _, attachedNode := getSv4TestAppVol(ctx)
-					Step(fmt.Sprintf("remove label on node for %s", ctx.App.Key), func() {
-						err := core.Instance().RemoveLabelOnNode(attachedNode.Name, "attachedNode")
-						Expect(err).NotTo(HaveOccurred())
-					})
-				}
-			})
-		})
-
 		AfterEach(func() {
 			// Do the cleanup in case the test failed before doing this.
 			Step("uncordon attachments on the other replica node", func() {
@@ -1236,17 +1117,6 @@ func getTestSv4Contexts(contexts []*scheduler.Context) []*scheduler.Context {
 		testSv4Contexts = append(testSv4Contexts, ctx)
 	}
 	return testSv4Contexts
-}
-
-func getTestSharedV4Contexts(contexts []*scheduler.Context) []*scheduler.Context {
-	var testSharedV4Contexts []*scheduler.Context
-	for _, ctx := range contexts {
-		if !strings.HasPrefix(ctx.App.Key, "test-sharedv4") {
-			continue
-		}
-		testSharedV4Contexts = append(testSharedV4Contexts, ctx)
-	}
-	return testSharedV4Contexts
 }
 
 // returns the appCounter structs for the app pods by scanning the export path on the NFS server
